@@ -1,5 +1,6 @@
 package de.dueto.backend.service;
 
+import de.dueto.backend.DatabaseTest;
 import de.dueto.backend.model.group.Group;
 import de.dueto.backend.model.user.User;
 import de.dueto.backend.mysql_data.UserRepository;
@@ -11,6 +12,7 @@ import org.springframework.context.ApplicationContextInitializer;
 import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.testcontainers.containers.MySQLContainer;
+import org.testcontainers.shaded.org.bouncycastle.asn1.dvcs.Data;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -18,25 +20,7 @@ import java.util.List;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 @SpringBootTest
-@ContextConfiguration(initializers = {de.dueto.backend.service.UserServiceTest.Initializer.class})
-class UserServiceTest {
-
-    public static MySQLContainer mySQLContainer = new MySQLContainer("mysql:8.0")
-            .withDatabaseName("DueTo-Database")
-            .withUsername("DueTo")
-            .withPassword("password");
-
-    static class Initializer
-            implements ApplicationContextInitializer<ConfigurableApplicationContext> {
-        public void initialize(ConfigurableApplicationContext configurableApplicationContext) {
-            mySQLContainer.start();
-            TestPropertyValues.of(
-                    "spring.datasource.url=" + mySQLContainer.getJdbcUrl(),
-                    "spring.datasource.username=" + mySQLContainer.getUsername(),
-                    "spring.datasource.password=" + mySQLContainer.getPassword()
-            ).applyTo(configurableApplicationContext.getEnvironment());
-        }
-    }
+class UserServiceTest extends DatabaseTest {
 
     @Autowired
     UserRepository userRepository;
